requirements.txt
asgiref==3.4.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Deprecated==1.2.13
Django==4.0
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
finnhub-python==2.4.7
idna==3.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
requests==2.27.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
wrapt==1.13.3
